use telnet-ssl instead of telnet as dependency inside the ngcp-nth-* packages
In today's dev meeting we noticed that telnet is missing on our
pro installation which is annoying. We already have a depency on
telnet in current trunk but it's not in the 2.3 repository yet,
that's why it was missing in today's pro installation. While at
it replace dependency on telnet with telnet-ssl since Daniel
suggested that.

From: Michael Prokop <mprokop@sipwise.com>