MT#7335 Accept Debian/jessie's heartbeat package as alternative to heartbeat-2
The heartbeat-2 package we use is based on version 2.1.3-7,
dating back to 2008 in Debian.

In Debian wheezy the heartbeat package has version 1:3.0.5-3,
and in jessie it's currently sitting at 1:3.0.5+hg12629-1.

We should investigate whether the heartbeat package from
Debian/jessie might work for us, so accept it as an alternative
to our heartbeat-2 package dependency-wise. We then can try
building a ngcp system based on Debian/jessie and if the
heartbeat 3.* package doesn't work for us we can still use our
own package (and optionally also revert this commit).