MT#8961 Ship locales for supporting /etc/default/locale file
We might want to ship the locales package on *all* our systems by
just adding it to the default package list, but it adds 15.2MB of
additional disk space which might be unwanted for certain VMs and
I'm not yet sure whether it would be acceptable to add the
locales package also on older sip:provider releases. Until we've
decided whether that would make sense or not let's install the
locales package via the ngcp-nth-{ce,pro} metapackage in our
sip:provider systems.

The file /etc/default/locale gets generated by the locales
package:

| # cat /etc/default/locale
| # File generated by update-locale

and then the tons of:

| Unable to open env file: /etc/default/locale: No such file or directory

in /var/log/auth.log disappears.

By also executing:

  update-locale "LANG=en_US.utf8"

we would get:

| LANG=en_US.utf8

added to the file, which might also make sense.

Change-Id: I1aaf2238caba10e5d10fc006e032116c127172ac