Update patch set 3
Patch Set 3:

> BTW, it means mytop will not be available for mr5.4+ (which will be
 > jessie based with a high probability).... :-(

mytop is in the nth (nice to have) package list, so not strictly a necessary tool (and BTW, is it even used by any ops/support/devs "kind of regularly"?).

If mytop should be installed anyway then we could maybe use something like "mariadb-client-10.1 | mytop" in Depends (and differencing between master vs stretch/master would be another option, but I'd avoid that as long as possible), but I'm not sure it's worth it?

Patch-set: 3