Update patch set 1
Patch Set 1:

As I've mentioned to Roman on jabber, the sems-pbx needs to grow some Replaces/Conflicts before this can happen cleanly. Also it might make sense to turn the sems-pbx into ngcp-sems, so that we have a less confusing package name? ideally we'd also be able to use that on the sems git repo, but oh well. :)

Patch-set: 1