Update patch set 1
Patch Set 1:

> I don't strongly object, though it feels wrong to me to add
 > dependencies on all those packages here. Some of them don't seem
 > relevant to our platform. I think we should discuss how all those
 > dependencies should be handled in grml-debootstrap/deployment.sh vs
 > installer vs metapackages etc

Ah definitely, sorry I didn't make this clear here, this was mostly to start a discussion and to track those packages, I think I mentioned this to Alex on jabber. The point was to list the current set of packages that when marked as automatic would get removed. There are even two firmware packages that are not listed here.

Patch-set: 1
Label: Code-Review=0