Update patch set 1
Patch Set 1:

I would like to assess where we are with this, where the stoppers are and if it works in general for a carrier, then optimize the templates for a PRO. I'd say this will take a few days.

I know there is lack of a tool to actually migrate kamailio.location, I'd like to discuss this with Kirill once the general setup works.

There are still some unknowns here (e.g. what happens or should happen if a new pair of proxies is added afterwards). Therefore, I'm not confident this is going to land in 5.5, but in any case this must be picked up and finalized now, no matter if the result will be in 5.5 or 6.0.

Patch-set: 1
Label: Code-Review=0