TT#24405 Install ngcp-installer and ngcp-upgrade on CE/PRO/Carrier by default
We are going to add new scripts into ngcp-upgrade-* packages to allow
the support team prepare upgrade better, like check list of supported release
and perform the pre-upgrade checks. It means the package must be installed out of the box.

Also good to ensure proper ngcp-installer-* is also installed as it is necessary for iPXE.
The package ngcp-installer was installed only because user/deployment.sh did it manually.

Maybe ngcp-nth-* is better place for such packages, but
we have to add them into ngcp-service-* as ngcp-nth-carrier doesn't
exist and list of added packages is different for PRO/Carrier.

+ wrap-and-sort -sat

Change-Id: I541a6f81e730327ef39913a00088eb052b5031ea