Update patch set 1
Patch Set 1:

> As I mentioned on jabber, I'd like to have it, but every time I've
 > tried it, I've ended up hating it. It breaks some normal stuff for
 > me, which is maddening. I never completely remember, and I should
 > file a bug report, but I think it's something related to standard
 > tab-completion?

TBH, I also don't use it on my systems because for example you want to autocomplete with a filename and if the completion is not well programmed (or up-to-date) for that command, it will refuse to do this most-basic of acompletions, so it's very annoying unless the support is complete and completely out of date.

(It happens for other things that are not filenames, but I rarely autocomplete @hostnames or whatever).

I +2'ed it anyway because I assume that many people cannot live without it, and I can live with it but not that I am a huge fan :)

Patch-set: 1