Update patch set 1
Patch Set 1: Code-Review+2

> My $0.05 cent here, I would give it a try at lest while we in green
 > zone for mr6.2 if we will see some issues here we can revert it
 > without releasing. Mika, your's final vote here :-)

I'm a Zsh user with a very strong completion (way beyond what bash does), I'm all for it if it's working fine. Since it seems to be active by default only for the sipwise user but not for the root user I think it's OK to just give it a shot. What might be a good idea is adjusting /var/sipwise/.bashrc by giving users an option to disable just bash-completion but keep the rest of the bashrc available (`bash --norc` provides a different environment and there doesn't seem to be a way how to disable completion once it's active already).

Patch-set: 1
Reviewer: Michael Prokop <1000000@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2