Update patch set 4
Patch Set 4: Code-Review+2

Self-approving because now it's implemented as feedback in other changes requested, no feedback for a week now (except today Mykola with +1, thanks!).

There's plenty of time to refine, revert or whatever is necessary before the next release, in the case that something is wrong.

Patch-set: 4
Label: Code-Review=+2