Update patch set 1
Patch Set 1: Code-Review+2

This metapackages is independent issue:

10:43:17    ngcp-hardware-support : Depends: firmware-bnx2 but it is not installable
10:43:17                            Depends: firmware-bnx2x but it is not installable

First of all it is the latest piuparts which Jenkins ignores...
We were not able to make it completely green so far,
this is why firmware issue was not noticed there. I believe it happens because non-free repository is disabled on piuparts. Separate ticket should be fired and addressed here IMHO.


LGTM for the review.

Patch-set: 1
Reviewer: Alexander Lutay <1000001@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+2