Update patch set 1
Patch Set 1:

> tl;dr: some packages are inherently tied to the hardware where the systems are installed

We no longer know the type of hardware. Carrier != IBM anymore.
We have carrier nodes on Dell nowadays and more are coming. Visualization?

> and we cannot completely avoid the problem in some cases that some hardware needs a different set of packages, possibly incompatible with other hardware in which we install this software.

Sure, in this case we have better notice it as soon as possible and fix/workaround.  Otherwise we will be blind for a year and huge Carrier customer will be crashed on upgrade, so far 100500 CE/PRO systemd were upgraded without the issues.

If driver conflicts - it must be fixed.
If driver install and no hardware availabele - it should politely sleep.
if DKMS upgrade fails - we should notice and fix it ASAP, without waiting for Carrier crash.

Carrier brings the money to Sipwise... it has biggest priority in stability.

Patch-set: 1