Update patch set 1
Patch Set 1:

> > tl;dr: some packages are inherently tied to the hardware where
 > the systems are installed
 > 
 > We no longer know the type of hardware. Carrier != IBM anymore.
 > We have carrier nodes on Dell nowadays and more are coming.
 > Visualization?

Precisely that's why I think that might be a problem.  For example, when searching for "megaraid-sas" there are different versions to download from Lenovo and Dell, the hardware is from Broadcom, and the driver is available in the stock kernel from Debian (yet it doesn't work in these Lenovo systems).

So it's not unthinkable that the Carrier systems installed on Dell that use this megaraid-sas work well with the megaraid-sas driver from the stock kernel but will not work (or is slower, or corrupts data) with our new "ngcp-megaraid-sas" that was taken from Lenovo, or that we need both "ngcp-megaraid-sas-dell" and "ngcp-megaraid-sas-lenovo" but only one can be installed at a time in the system.

Having an Intel eth card that doesn't work is bad, but I am more worried about the worse case that disk-drivers that we add take over the previous driver in systems that do not need it and that we could have avoided installing the package, and corrupt data or something bad.


 > > and we cannot completely avoid the problem in some cases that
 > some hardware needs a different set of packages, possibly
 > incompatible with other hardware in which we install this software.
 > 
 > Sure, in this case we have better notice it as soon as possible and
 > fix/workaround.  Otherwise we will be blind for a year and huge
 > Carrier customer will be crashed on upgrade, so far 100500 CE/PRO
 > systemd were upgraded without the issues.
 
Hmm, where did you get the number of 100500?  Do we have "telemetry" data? :-)


 > If driver conflicts - it must be fixed.
 > If driver install and no hardware availabele - it should politely
 > sleep.
 > if DKMS upgrade fails - we should notice and fix it ASAP, without
 > waiting for Carrier crash.

So if we have telemetry that confirms that at least installs and works well in the systems that are being upgraded, it'll be a step forward.

As I said above I am more concerned with drivers that can cause not only crashes but potentially more harm, possibly irrecoverably, like the megaraid-sas that we're also testing for new Carrier systems.

So it's not only a case of CE/Pro/Carrier, but differences within Carrier systems that might cause problems when installing drivers.

...

That said, do you want to go ahead with installing ngcp-intel-i40e in all systems?

Patch-set: 1
Label: Code-Review=0