Update patch set 1
Patch Set 1:

> Please mention the relevant Debian bug number in the commit
 > message, so we can track this down in the future.

That would be https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=875708 , but that's in stable and unfixed for 1+ years there, so it's even likely that it will not be fixed for stretch.

However, it's fixed already for "buster". So probably we'll not be able to take it off of metapackages until we move away from stretch into buster.

But then there's a different one  that you found for "buster", as I mentioned in the ticket in workfront ( https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=869782 ), which needs other packages installed, and that one unfixed in "buster" for ~1 year too.

Patch-set: 1
Label: Code-Review=0