TT#49026 Add snmptrapd to the default metapackages
We have configuration setup for snmptrapd to help debugging or for local
triggers. But if the local admin enables it, the package will be missing.
Let's install it by default, because it's a very tiny binary, which will
not be started as long as it's not enabled.

Change-Id: Ibe3952d9b017637cf7bd05a51558d6922ba74f78
(cherry picked from commit 12068f28196c3c1bd9fae5bfec9c14e5e37ff38e)