TT#63655 Switch to depend on libngcp-bulkprocessor-projects-perl on PRO
This module got split out from libngcp-bulkprocessor-perl, as some of
the actual project implementation modules, with tools, which act as
the frontend.

Eventually the tools might get split out into their own package, but
for now we should depend on the new package on PRO which is the one
that provides some of the actual end-user functionality.

Change-Id: I54bbb9ac22a306c2da2040badee4bdecfc966257
(cherry picked from commit 3a4a8ef8efb71af16fc0096f679a812508343ede)