TT#63655 Switch to depend on libngcp-bulkprocessor-projects-perl
This module got split out from libngcp-bulkprocessor-perl, as the
actual project implementation modules, with tools, which act as the
frontend.

Eventually the tools might get split out into their own package, but
for now we should depend on the new package which is the one that
provides the actual end-user functionality.

Change-Id: I54bbb9ac22a306c2da2040badee4bdecfc966257