TT#70750 Add support for Linux cloud kernels
We need to add a metapackage indirection so that we can select the image
and headers packages that match.

This is a problem on cloud systems, were the installed and running kernel
is a cloud variant, but we are pulling in a non-cloud image and headers,
which will be used by dkms when building kernel modules, but not for the
running one.

By adding this indirection we make specify that we need both the image
and the headers to match, preferring the stock versions, but selecting
the cloud headers if such kernel image is already installed.

Change-Id: I85e76ff56e6796c4315ebe6bef2e19e63cb7b362