Update patch set 1
Patch Set 1:

I think that the problem is that it doesn't install in the given environment, because when attempting to install the kernel module it attempts to do "strange" things (from a "cloud-y" point of view) of inserting modules in the kernel and so on.

The kernel driver package is built using the standard way in Debian to build kernel modules, and the result works well in real hardware.

So I think that probably we can just ignore piuparts in this case, and move on.

Patch-set: 1
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>