Update patch set 2
Patch Set 2:

> Patch Set 2:
> > which seems unrelated to this package.  So yeah, I suppose that the specifics of the testing environment (docker?) is what makes this module fail to build, it works fine in regular chroots.
> 
> We're running piuparts inside VMs, looking like regular systems, your build seems to be working though. What piuparts version are you using?

  0m0.0s INFO: piuparts version 1.0.0+deb10u1 starting up.

I installed in our very-own ce-trunk.mgm.sipwise.com, and it's a Debian-based chroot and not a Sipwise-based one, but in principle it should be OK for this package (kernel driver, only depending on the kernel headers and compiler and little else).

Note that the failure in this review with piuparts is in the metapackages and not the kernel driver, but due to the chroot being based on Debian, I didn't check with the metapackages in my test (I'm not sure how to have a sipwise-based chroot, for example our mirror was not working when telling debootstrap to use it, which I attempted at first).

But the cause of the failure of the metapackages is when building the kernel driver package, so I think that my test is relevant to show that there's nothing wrong with the package itself (it got built and passed its own piuparts tests last week, btw).

But yeah, if we're using a regular VM for piuparts, not sure why it's not building fine there, ce-trunk.mgm.sipwise.com is also a VM after all...

Patch-set: 2
Reviewer: Gerrit User 1000042 <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>