Update patch set 2
Patch Set 2:

> LGTM, just in case, I am sure you did, but...
 > Did you consider upgrade mr7.5->mr8.5 LTS process when inactive A
 > is in new release and active B is still old.
 > So some wired cases we can stay for a day (like was on Netcologne
 > and another customer once).
 > 
 > Will new inactive A node continue working using old monitoring and
 > old active B also work well?
 > Just to simulate such situation in a mind. Sure B monitoring will
 > know nothing about A node and will report it as down...

JFYI this commit does not affect the upgrade stuff, as mentioned on the commit message the dependency is present in the check-tools package for now. This will be relevant with the WIP commit removing that from check-tools. 

But I've not started the migration part yet. But I've been testing my changes with a node converted to prometheus and no influx running, and a node still using influx. And I don't see any major problems, the metrics will just be missing from one side. My idea is to remove everything except influxdb, until both have been switched so that grafana can still work, but of course things like collective-check will start to trigger, so that might be an issue dunno. :/

Patch-set: 2
Reviewer: Gerrit User 1000016 <1000016@85315e57-5f03-49c3-83f8-201ae2313a75>