TT#80304 Install ngcp-admin-ui by default
We have received a green light for ngcp-admin-ui from CTO,
we are going to make it reachable for mr8.5+ to gather the
very initial customer's feedback and correct our vision/plans if necessary.

ngcp-admin-ui will be disabled by default in mr8.5,
it will be enabled for the necessary installations only.

It is safe to install the package right now from the security point of view.

Also "wrap-and-sort -sat" list of dependencies here.

Change-Id: I9c69982799ed776eb8188bd0ea4d018441a34db6