TT#112302 Move monitoring stack daemons dependencies from ngcp-check-tools
The dependencies should be in the metapackage, as there's no dependency
at all for any of the Prometheus exporters, and there's just an indirect
dependency on the database services, but we do not usually list them as
dependencies anyway (f.ex. mariadb) in ngcp-check-tools. This also makes
it possible to not depend on PRO/Carrier specific packages on CE systems.

Change-Id: I68e0d3fc26753c629712d94c85fc3d76ed212ae7