TT#144703 Restore support for ntp
We add optional support for ntpd, as some customers have demanded it
back. While preserving timesyncd as the default.

We need to add ntp as an alternative dependency to systemd-timesyncd,
although our policy is to install all services (even when disabled),
so that the service can be enabled or switched just via configuration
changes. But in this case all time-servers Conflicts/Replaces with each
other, so we cannot have them all installed, and that will require
administrator intervention.

Change-Id: I75c25010c3c12dfeda0c0297c7cea57076ae1bf5