TT#160551 Add percona-toolkit to the addons packages
We depended on this from the ngcp-check-tools package, as that was a
direct user of these tools. But since we have removed explicit checks
for database replication discrepancies, there's no use for that there
anymore. But these tools are useful for diagnosing database problems,
so let's move the dependencies into the addons packages.

Change-Id: Id6d66330424cd24d3e199afc2d8955e8c1ba3ba3