MT#55468 keydb-server < 6.3
Last available version of keydb: 6.3.1-1~bpo11+1 does not look stable.
Many issues have been reported already in customer systems and also
a lot of issues are now opened in the upstream project.

To be on the safe side, better to postpone the upgrade of keydb
until the new version will be more stable and mature.

Change-Id: Ife3115b7f7d1f4b331f55c773323d15cece340c8