MT#56541 Add yq to ngcp-addons-{ce,pro}
yq is a jq wrapper for YAML documents and is useful when having to deal
with YAML files as used e.g. in cfg-schema. As of bookworm, this package
is provided by Debian.

FTR: the tool requires 854 kB of additional disk space on CE and 349 kB
of additional disk space on PRO.

Change-Id: Ifbbe61186effca008d02f477c509d9900eb7a678