MT#61101 Install valkey-server in ngcp-services-*
We are introducing support for Valkey as an alternative to KeyDB (in
itself an alternative to Redis, which now has the additional problem
of its non-free license), which we are intending to switch on all
installations that do not rely on KeyDB specific features currently
not present in Valkey.

KeyDB has an uncertain future, being managed and developer by a single
vendor (Snapchat) who has stopped interacting publicly on its GitHub
project, where its primary public face developers has quit the company,
and where Snapchat is part of the corporate supporters for Valkey.

So we better start planning our escape route, before we are left with
a potentially unmaintained project that no one else is using or
supporting.

Change-Id: Idf5ab9dac4e085070dbdc6d63cddfeeaf495ca9f