MT#63650 Switch NTP default backend to ntpsec
For PRO/Carrier where this setting makes most sense, the
systemd-timesyncd service does not provide accurate nor reliable
fallback.

Having to support both implementations when we are required to support a
proper one, which will be the default going forward means more testing
surface for not much benefit, so we switch completely.

Change-Id: I6349c5f109530040ae010725daac87441a0e7794